Update audio_policy.conf
Fixes audio pops and crackles when listening to music. I have tested for a few hours on spotify as well as multiple users on telegram tested and stated it fixed the audio defects, and were also able to run both core clusters at 200mhz min. Thus able to revert change 163214. However, the crackles in notifications still continue with or without this edit.